Improve reliability of cancelling resuming.
Improve the reliability of cancelling resuming by ensuring that
we use the copy of the image header on disk post-atomic-restore
(and not the version from when there was no image). In addition,
make sure that we clean up the modules before trying to rewrite
the signature, so as to avoid low-memory situations causing a
failure to allocate a page for the signature and then an oops
prior to rewriting the signature.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>