Add support for scanning for an image.
Add support for searching for an image if resume= is blank. This
initial version extends the uuid code a little to enable finding
partitions of a particular type. We use this code for swap,
checking each swap partition found for an image header. The first
partition found is remembered, and becomes a default resume=
value if not image is found. If an image is found, that partition
is used for resume= instead.

The end result is that users should now be able to hibernate and
resume without ever having to worry about setting resume=, if
they have at least one swap partition. Even if that swap partition
isn't used for hibernating, the cycle will still success (as it
did before), because resume= just gives the location of the start
of the image header, which may be on a completely different device
again.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>