Fix storage allocation mistake.
While reading the code, I noticed a mistake in the storage allocation
loop. We should be looping over all the storage devices of a priority,
not some potentially lower number. This mistake would show up (for
example)in a scenario where there is less storage available on the
first device of a priority than the second, and multiple passes of
the loop are required to obtain the amount of storage desired. If
we ran out of storage on the first device on the first pass, it would
be the only device from which we attempt to allocate storage on the
second pass, and we would move on to lower priority storage (if any)
with storage  on the second device still available.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>