Correct calculation of throughput throttle.
596fee0523a569c261d73980943bf1ca29e4b718 ("Give an estimate of time
remaining") adjusted the calculation of the throughput throttle
because we're now updating the throttle each 1/4 of a second instead
of 1/2 second. The change wrongly switched a divide to a multiply,
causing us to have much more outstanding I/O than is necessary and
enough such that our remaining time calculations are off. Switch
it back to a divide.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>