Fix kmap while atomic warning (speedup!)
Pedro Ribeiro reported that TuxOnIce was triggering a WARN_ON at
line 380 in kernel/smp.c, because it was calling kmap while in
an atommic context.

This was happening in the code that decides where pageset1 will be
loaded, because we were allocating all available memory as part of
that process, and going atomic to avoid other code causing OOM
complaints.

It should have occured to me long ago that there's no need to allocate
all available memory. We can just allocate the number of pages we
need, then treat the ones that match with pages to be restored as
direct loads, and the others as source pages for the atomic restore.

We gain a nice speed improvement this way, too. We (at least potentially)
do far fewer memory allocations and frees (especially on systems with
lots of memory and a comparatively small pageset1), and we completely
avoid the cost of disabling and reenabling secondary processors.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>