Add image header versioning.
Add a version number to the image header, so that we can avoid future
issues with backwards compatability. Since previous versions haven't
had a version number and the bytes we're going to use haven't been
overwritten, use an additional copy of the binary signature so we can
be sure we really have a version number and not whatever might have
been in those bytes before the page was used for the image signature.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>