Allow the user to limit the number of threads used.
Add support for limiting the number of threads we use for
doing I/O, providing a new sysfs entry that can be used
to set the value. The default of zero provides the old
behaviour - one per cpu.

The impetus for this patch is the report from Nix that
on his eight core box, throughput was much lower than
expected. Let's see if it's due to lock contention.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>