Fix unitialised output size value.
LZO expects output_size to be initialised to the amount of space
available in the output buffer prior to calling its decompression
routine. Fix this in the compression checking routine (it's already
initialised in the main version) so as to avoid false failure
reports such as the one I passed on to Richard and Nitin.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>