Fix handling of no data left to read case.
TuxOnIce has occasionally been reported to return -ENODATA at the
end of resuming, resulting in an oops. I've discovered today that
the cause is actually I/O threads not getting run until all the
data has been read. In this situation, they try to read bogus data
on their first iteration through the main loop. The solution is to
check the count of outstanding I/O up front, bypassing the plugins
if there's no data to read, and dropping straight through to the
cleanup.

As part of this, we need to adjust the error handling in the main
loop so that io_result isn't to -ENODATA when this happens, which
will also correct the handling of the situation in the case where
we're writing the image.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>