Set num workers before starting threads.
Rework the handling of toi_io_workers so that instead of
incrementing it as each thread starts, the value is set
prior to the threads being started. This helps ensure that
we won't call the last_worker logic (at the bottom of the
loop code) multiple times if one or more threads don't
end up doing any work.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>