Better handling of duplicated UUIDs.
Improve our handling of the situation where a user utilises dd
to copy a partition, and doesn't change the UUID on the destination.
Do this by using not just the UUID, but also the last mounted time
and dev_t to try to match partitions, recording all of these pieces
of information in image header when hibernating, and picking the
best match at resume time. "Best" is defined as having all values
matching in the ideal case, taking a partition with a matching
UUID and last mount time as the next best match, and only falling
back to taking the first UUID match if nothing better is found.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>