Hibernation: Check all partitions for fs info
There was one more bug in the blk_lookup_fs_info code: it was
still dropping out prematurely. In the case where a person copies
a partition from an ordinary partition into an LVM volume, we
need to check multiple block classes to ensure we find the right
partition. In fact, we should always check all block classes and
all partitions (unless we get the best possible match first).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>