irq: implement IRQ watching
This patch implements IRQ watching, which is a simple polling
mechanism drivers can use to work around lost and/or misrouted IRQs.
IRQ watching is enabled by driver calling watch_irq(irq, dev_id).
After that, it polls the irqaction for certain amount of time (1min)
and keeps track of whether IRQ delivery is actually working.  If the
irqaction is serviced by poll, it's considered to be a possible
indication of IRQ misdelivery.

The watch polling starts slowly at 1HZ and speeds up to 100HZ when it
sees a possible bad delivery.  After collecting stats for a while, it
determines whether the IRQ delivery is working for the irqaction.  If
so or it can't be determined, it steps out.  If not working, IRQ
polling continues till the irqaction is unregistered.

This can be used by drivers which don't know when to expect the next
IRQ.  Just by calling watch_irq() after irqaction is registered or a
timeout, most IRQ misrouting problems can be worked around.

Signed-off-by: Tejun Heo <tj@kernel.org>