cfq: always return false from should_idle if slice_idle is set to zero
Hi,

In testing a competing fsync-ing process and a sequential reader on
mid-grade storage, I found that cfq was incapable of achieving the I/O
rates of deadline, no matter how it was tuned.  Investigation, and insight
from Vivek (mostly the latter), led to identifying that we were still
idling for the last queue in the service tree.

Modifying cfq_should_idle to not idle when slice_idle is set to zero got
us much closer to the performance of deadline for this workload.  I have
one follow-on patch that gets us on-par with deadline, but I think this
patch stands alone.

Comments, as always, are appreciated.

Cheers,
Jeff

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>