cfq: allow dispatching of both sync and async I/O together
Hi,

In testing a workload that has a single fsync-ing process and another
process that does a sequential buffered read, I was unable to tune CFQ
to reach the throughput of deadline.  This patch, along with the previous
one, brought CFQ in line with deadline when setting slice_idle to 0.

I'm not sure what the original reason for not allowing sync and async
I/O to be dispatched together was.  If there is a workload I should be
testing that shows the inherent problems of this, please point me at it
and I will resume testing.  Until and unless that workload is identified,
please consider applying this patch.

Cheers,
Jeff

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>