TuxOnIce: Retry opening secondary storage
I received a report:

http://bugzilla.tuxonice.net/show_bug.cgi?id=458

that TuxOnIce was failing to wait for all hard drives to be
found. The problem was that the logic that waits for async
scans and retries opening devices had only been applied to
the device used for resume=, and not any other devices used
in the image.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>