Improve handling at end of reading data.
Improve the handling when finishing reading data. Particularly in
multithreaded operation, there are lots of possible places that
each thread can be in. Most of these changes were done while working
on the code that tried making everything lockless. That code won't
be merged (it actually slowed things down), but this is some fruit
that will see the light of day.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>