Hold secondary threads until setup is done
This patch fixes a bug reported on the tuxonice-devel mailing
list, in which TuxOnIce has an oops almost as soon as it starts
writing the image. The bug is due to secondary threads being
allowed to start running immediately, rather than being made to
wait until the first thread has finished setting up. Address this
by having the first thread take the io_mutex prior to starting
secondary threads, and drop it before entering the io flusher or
worker loop itself.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>